% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeTracksWidget.R
\name{GenomeTracksWidget}
\alias{GenomeTracksWidget}
\alias{printf}
\title{GenomeTracksWidget}
\usage{
printf(...)
}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing
}
\description{
an R6 class providing clean access to Gene Ontology Enrichment, input & display
}
\details{
R6 Class representing one or more igv genome tracks, from tf/targetGene data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenomeTracksWidget$new()}}
\item \href{#method-ui}{\code{GenomeTracksWidget$ui()}}
\item \href{#method-server}{\code{GenomeTracksWidget$server()}}
\item \href{#method-addAllRelevantTracks}{\code{GenomeTracksWidget$addAllRelevantTracks()}}
\item \href{#method-clone}{\code{GenomeTracksWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new GenomeTracksWidget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeTracksWidget$new(id, title, tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{title}}{character}

\item{\code{geneSymbols}}{character vector, HUGO standard names only}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `GenomeTracksWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeTracksWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeTracksWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addAllRelevantTracks"></a>}}
\if{latex}{\out{\hypertarget{method-addAllRelevantTracks}{}}}
\subsection{Method \code{addAllRelevantTracks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeTracksWidget$addAllRelevantTracks(gene, newTrackWidget)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeTracksWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
